+++ eslint-config-dsx-vue/rules/typescript.js
@@ -6,14 +6,14 @@ module.exports = {
   ],
   settings: {
     'import/parsers': {
-      '@typescript-eslint/parser': ['.ts','.tsx']
-     },
+      '@typescript-eslint/parser': ['.ts', '.tsx'],
+    },
   },
   plugins: ['@typescript-eslint'],
   parser: 'vue-eslint-parser',
   parserOptions: {
     parser: '@typescript-eslint/parser',
-    project: ['./tsconfig.json']
+    project: ['./tsconfig.json'],
   },
   rules: {
     '@typescript-eslint/no-var-requires': 'warn',